+++ front/documents_post_process_hooks/temporal/worker.ts
@@ -2,7 +2,7 @@ import { Context } from "@temporalio/activity";
 import { Worker } from "@temporalio/worker";
 
 import * as activities from "@app/documents_post_process_hooks/temporal/activities";
-import { getTemporalWorkerConnection } from "@app/documents_post_process_hooks/temporal/lib";
+import { getTemporalWorkerConnection } from "@app/lib/temporal";
 import { ActivityInboundLogInterceptor } from "@app/lib/temporal_monitoring";
 import logger from "@app/logger/logger";
